feat(int64): rename and convert helper files to long indexing
File renames:
- MemoryExtensions*.cs → UnmanagedSpanExtensions*.cs
- Buffer.cs → UnmanagedBuffer.cs

Class renames:
- MemoryExtensions → UnmanagedSpanExtensions
- Buffer → UnmanagedBuffer

int → long conversions across all helper files:
- Method parameters: length, searchSpaceLength, valueLength, start
- Return types: IndexOf, LastIndexOf, Count, BinarySearch, etc.
- Local variables: index, i, offset

Files converted:
- UnmanagedSpanHelpers.cs (base helpers - already used nuint)
- UnmanagedSpanHelpers.T.cs (generic IndexOf, Contains, SequenceEqual)
- UnmanagedSpanHelpers.Byte.cs (byte-optimized operations)
- UnmanagedSpanHelpers.Char.cs (char-optimized operations)
- UnmanagedSpanExtensions.cs (main extension methods)
- UnmanagedBuffer.cs (Memmove - already used nuint)

This enables all span operations to work with >2 billion elements.

+++ src/NumSharp.Core/Utilities/SpanSource/UnmanagedSpanHelpers.BinarySearch.cs
@@ -21,7 +21,7 @@ namespace System
         }
 
         public static int BinarySearch<T, TComparable>(
-            ref T spanStart, int length, TComparable comparable)
+            ref T spanStart, long length, TComparable comparable)
             where TComparable : IComparable<T>, allows ref struct
         {
             int lo = 0;
@@ -34,8 +34,8 @@ namespace System
                 //       `length <= int.MaxValue`, and indices are >= 0
                 //       and thus cannot overflow an uint.
                 //       Saves one subtraction per loop compared to
-                //       `int i = lo + ((hi - lo) >> 1);`
-                int i = (int)(((uint)hi + (uint)lo) >> 1);
+                //       `long i = lo + ((hi - lo) >> 1);`
+                long i = (int)(((uint)hi + (uint)lo) >> 1);
 
                 int c = comparable.CompareTo(Unsafe.Add(ref spanStart, i));
                 if (c == 0)
